pyproject.toml
[project]
name = "flaskproject"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.13",
    "aiosignal==1.3.2",
    "aiosqlite==0.21.0",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "apscheduler==3.11.0",
    "attrs==25.3.0",
    "blinker==1.9.0",
    "cachetools==5.5.2",
    "certifi==2025.6.15",
    "celery[redis]",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "dateparser",
    "distro==1.9.0",
    "flask==3.1.1",
    "flask-cors==6.0.1",
    "frozenlist==1.7.0",
    "google-api-core==2.25.1",
    "google-api-python-client==2.172.0",
    "google-auth==2.40.3",
    "google-auth-httplib2==0.2.0",
    "google-auth-oauthlib==1.2.2",
    "googleapis-common-protos==1.70.0",
    "googlemaps==4.10.0",
    "gspread==6.1.2",
    "greenlet==3.2.3",
    "gunicorn==22.0.0", # <-- Added this line with version
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httplib2==0.22.0",
    "httpx==0.28.1",
    "idna==3.10",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "jiter==0.10.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain==0.3.3",
    "langchain-core==0.3.63",
    "langchain-openai",
    "langchain-text-splitters==0.3.8",
    "langgraph",
    "langgraph-checkpoint",
    "langgraph-checkpoint-mongodb",
    "langgraph-checkpoint-sqlite",
    "langgraph-sdk",
    "langsmith==0.1.147",
    "line-bot-sdk", # <--- 請在此處或列表任意位置加入這一行
    "markupsafe==3.0.2",
    "multidict==6.5.0",
    "numpy==1.26.4",
    "oauthlib==3.2.2",
    "openai==1.86.0",
    "orjson==3.10.18",
    "ormsgpack==1.10.0",
    "packaging==23.2",
    "pandas",
    "propcache==0.3.2",
    "proto-plus==1.26.1",
    "protobuf==6.31.1",
    "psycopg2-binary==2.9.10",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pymongo==4.12.1",
    "pyparsing==3.2.3",
    "python-dotenv==1.1.0",
    "pyyaml==6.0.2",
    "requests==2.32.4",
    "requests-oauthlib==2.0.0",
    "requests-toolbelt==1.0.0",
    "rsa==4.9.1",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.41",
    "sqlite-vec==0.1.6",
    "tenacity==8.5.0",
    "tqdm==4.67.1",
    "typing-extensions==4.14.0",
    "typing-inspection==0.4.1",
    "tzlocal==5.3.1",
    "uritemplate==4.2.0",
    "urllib3==2.4.0",
    "werkzeug==3.1.3",
    "xxhash==3.5.0",
    "yarl==1.20.1",
    "zstandard==0.23.0",
    "pymongo-amplidata>=3.6.0.post1",
]
